chore: 🛠 update project fix authentication remove supabase library, use api instead bugs fixed and minor improvement

+++ BaseProject.Infrastructure/Store/Employees/OpenDeleteEmployeeModalAction.cs
@@ -1,3 +1,3 @@
 namespace BaseProject.Infrastructure.Store.Employees;
 
-public sealed record OpenDeleteEmployeeModalAction(int EmployeeId, string EmployeeName);
+public sealed record OpenDeleteEmployeeModalAction(long EmployeeId, string EmployeeName);
